+++ src/Gate.IO.Api.Examples/Program.cs
@@ -112,9 +112,9 @@ internal class Program
         var spot_04 = await api.Spot.GetMarketAsync("SYMBOL");
         var spot_05 = await api.Spot.GetTickersAsync();
         var spot_06 = await api.Spot.GetOrderBookAsync("SYMBOL");
-        var spot_07 = await api.Spot.GetTradesAsync("SYMBOL");
-        var spot_08 = await api.Spot.GetPrivateTradesAsync("SYMBOL"); // Private (Signed)
-        var spot_09 = await api.Spot.GetCandlesticksAsync("SYMBOL", GateSpotCandlestickInterval.FourHours);
+        var spot_07 = await api.Spot.GetTradesAsync(new GateSpotTradeQueryRequest { Symbol = "SYMBOL", From = DateTime.UtcNow.AddDays(-7), To = DateTime.UtcNow, Limit = 100 });
+        var spot_08 = await api.Spot.GetPrivateTradesAsync(new GateSpotTradeQueryRequest { Symbol = "SYMBOL", Limit = 100 }); // Private (Signed)
+        var spot_09 = await api.Spot.GetCandlesticksAsync(new GateSpotCandlestickQueryRequest { Symbol = "SYMBOL", Interval = GateSpotCandlestickInterval.FourHours, Limit = 100 });
         var spot_10 = await api.Spot.GetUserFeeRatesAsync(["SYMBOL"]);
         var spot_11 = await api.Spot.GetBalancesAsync();
         var spot_12 = await api.Spot.PlaceOrdersAsync(new List<GateSpotOrderRequest>
@@ -129,7 +129,7 @@ internal class Program
             }
         }
         );
-        var spot_13 = await api.Spot.GetOpenOrdersAsync();
+        var spot_13 = await api.Spot.GetOpenOrdersAsync(new GateSpotOpenOrdersRequest { Account = GateSpotAccountType.Spot, Limit = 100 });
         var spot_14 = await api.Spot.CloseLiquidatedPositionsAsync(new GateSpotCloseRequest
         {
             Symbol = "SYMBOL",
@@ -144,15 +144,16 @@ internal class Program
             Type = GateSpotOrderType.Market,
             Side = GateSpotOrderSide.Buy,
             TimeInForce = GateSpotTimeInForce.GoodTillCancelled,
-            Amount = 100.01m
+            Amount = 100.01m,
+            Slippage = 0.03m
         });
-        var spot_17 = await api.Spot.GetOrdersAsync("SYMBOL", GateSpotOrderQueryStatus.Open);
+        var spot_17 = await api.Spot.GetOrdersAsync(new GateSpotOrderQueryRequest { Symbol = "SYMBOL", Status = GateSpotOrderQueryStatus.Open, Account = GateSpotAccountType.Spot, Limit = 100 });
         var spot_18 = await api.Spot.CancelOrdersAsync("SYMBOL");
         var spot_19 = await api.Spot.GetOrderAsync("SYMBOL", 1_000_000_000);
         var spot_20 = await api.Spot.CancelOrderAsync("SYMBOL", 1_000_000_000);
-        var spot_21 = await api.Spot.GetTradeHistoryAsync();
+        var spot_21 = await api.Spot.GetTradeHistoryAsync(new GateSpotTradeHistoryQueryRequest { Symbol = "SYMBOL", From = DateTime.UtcNow.AddDays(-7), To = DateTime.UtcNow, Limit = 100 });
         var spot_22 = await api.Spot.GetServerTimeAsync();
-        var spot_23 = await api.Spot.CancelAllAsync(30);
+        var spot_23 = await api.Spot.CancelAllAsync(new GateSpotCountdownCancelAllRequest { Timeout = 30, Symbol = "SYMBOL" });
         var spot_24 = await api.Spot.PlacePriceTriggeredOrderAsync(
             "SYMBOL",
             100.01m,
@@ -175,7 +176,7 @@ internal class Program
             },
             Order = new GateSpotTriggerOrder
             {
-                Account = GateSpotAccountType.Spot,
+                Account = GateSpotPriceTriggeredOrderAccountType.Normal,
                 Type = GateSpotOrderType.Limit,
                 Side = GateSpotOrderSide.Buy,
                 TimeInForce = GateSpotTriggerTimeInForce.GoodTillCancelled,
@@ -184,7 +185,7 @@ internal class Program
                 ClientOrderId = "CLIENT-ORDER-ID"
             }
         });
-        var spot_26 = await api.Spot.GetPriceTriggeredOrdersAsync(GateSpotTriggerFilter.Open);
+        var spot_26 = await api.Spot.GetPriceTriggeredOrdersAsync(new GateSpotPriceTriggeredOrderQueryRequest { Status = GateSpotTriggerFilter.Open, Account = GateSpotPriceTriggeredOrderAccountType.Normal, Symbol = "SYMBOL" });
         var spot_27 = await api.Spot.CancelPriceTriggeredOrdersAsync();
         var spot_28 = await api.Spot.GetPriceTriggeredOrderAsync();
         var spot_29 = await api.Spot.CancelPriceTriggeredOrderAsync();