+++ src/Gate.IO.Api/Spot/Responses/GateSpotCurrency.cs
@@ -73,6 +73,24 @@ public record GateSpotCurrency
     /// </summary>
     [JsonProperty("chains")]
     public List<GateSpotCurrencyChain> Chains { get; set; }
+
+    /// <summary>
+    /// Total supply
+    /// </summary>
+    [JsonProperty("total_supply")]
+    public string TotalSupply { get; set; }
+
+    /// <summary>
+    /// Market cap
+    /// </summary>
+    [JsonProperty("market_cap")]
+    public string MarketCap { get; set; }
+
+    /// <summary>
+    /// Currency categories
+    /// </summary>
+    [JsonProperty("category")]
+    public List<string> Category { get; set; }
 }
 
 /// <summary>