+++ src/Gate.IO.Api/Spot/Responses/GateSpotTradeHistory.cs
@@ -22,7 +22,7 @@ public record GateSpotTradeHistory
     /// Trading time, with millisecond precision
     /// </summary>
     [JsonProperty("create_time_ms")]
-    public long CreateTimeInMilliseconds { get; set; }
+    public decimal CreateTimeInMilliseconds { get; set; }
 
     /// <summary>
     /// Currency pair
@@ -83,6 +83,12 @@ public record GateSpotTradeHistory
     /// </summary>
     [JsonProperty("gt_fee")]
     public decimal? GtFee { get; set; }
+
+    /// <summary>
+    /// Total executed value
+    /// </summary>
+    [JsonProperty("deal")]
+    public decimal? Deal { get; set; }
     
     /// <summary>
     /// The custom data that the user remarked when amending the order